==> notebooks/05_visualize_training_data.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
features_and_targets = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_targets


features = features_and_targets.drop(columns=['target_rides_next_hour'])
targets = features_and_targets['target_rides_next_hour']


from src.plot import plot_one_sample
plot_one_sample(
    example_id = 4,
    features = features,
    targets = targets,
)
